Add logout functionality to profile page and change redirect links.
